Create local cache for door API to use.
To save on internet usage I designed a simple local cache of users that
the Door API now will use.
Every time a new player comes close to the sensor the code will save
their access level to the cache. When the computer restarts the cache
will be gone. This allows for changes in the database while still
easying off on the internet usage and the delay for connecting to the
database.